Update VerticalIndex to uint8 and adjust dubcol(), dufcol() & dundcol() parameters. Avoid implicit sign conversions

+++ ThrEd4/ThrEdTypes.h
@@ -19,8 +19,8 @@
 
 constexpr auto RES_SIZE    = 26;               // reserved for expansion in the ThrEd v1.0 header
 constexpr auto NAME_LEN    = 50;               // Length of the name fields in ThrEd headers
-constexpr auto COLOR_COUNT = 16U;              // Number of colors in arrays
-constexpr auto COLOR_MAX   = COLOR_COUNT - 1U; // max index in color arrays
+constexpr auto COLOR_COUNT = gsl::narrow_cast<uint8_t>(16U);              // Number of colors in arrays
+constexpr auto COLOR_MAX   = gsl::narrow_cast<uint8_t>(COLOR_COUNT - 1U); // max index in color arrays
 constexpr auto COLOR_BITS  = 0xfU;     // Number of bits to cover the number of the colors in arrays
 constexpr auto SRTIM       = 20000000; // sort time limit in 100 ns intervals
